+++ delethon/metadata.py
@@ -7,7 +7,7 @@ Defines delethon's metadata.
 _ = str
 
 NAME = 'delethon'
-VERSION = '0.1.0-alpha'
+VERSION = '0.1.1-alpha'
 DESCRIPTION = _('Delete Telegram messages based on Telethon.')
 AUTHOR = 'Bing Ling'
 AUTHOR_EMAIL = 'binglinggroup@outlook.com'